backend/requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.9
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
protobuf==3.20.3
pycparser==2.23
pynput==1.8.1
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
six==1.17.0
sounddevice==0.5.3
